Win - Changes

+++ windows/Ball.cpp
@@ -28,7 +28,7 @@ void Ball::draw() {
 void Ball::draw(float d) {
     BeginDrawing();
     {
-        DrawTextureEx(sprite, Vector2 {x, y}, d, 1.0f, WColor::getWhite());
+        DrawTextureEx(sprite, Vector2 {(float)x, (float)y}, d, 1.0f, WColor::getWhite());
     }
     EndDrawing();   
 }
